requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
deprecated==1.2.7
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Markdown==3.2
MarkupSafe==1.1.1
mistune==0.8.4
peewee==3.13.1
pygithub==1.45
pyjwt==1.7.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
